+++ td/telegram/StickerFormat.h
@@ -25,7 +25,7 @@ bool is_sticker_format_animated(StickerFormat sticker_format);
 
 bool is_sticker_format_vector(StickerFormat sticker_format);
 
-size_t get_max_sticker_file_size(StickerFormat sticker_format, bool for_thumbnail);
+int64 get_max_sticker_file_size(StickerFormat sticker_format, bool for_thumbnail);
 
 StringBuilder &operator<<(StringBuilder &string_builder, StickerFormat sticker_format);
